requirements.txt
aniso8601==1.0.0
cryptography==0.9.1
ddt==1.0.0
elasticsearch==1.6.0
Flask==0.10.1
Flask-RESTful==0.3.3
flask-restful-swagger==0.19
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
mock==1.0.1
PyJWT==1.3.0
pytz==2015.4
requests==2.7.0
six==1.9.0
urllib3==1.10.4
Werkzeug==0.10.4
